small fixes

+++ libs/event/src/lib/pages/login/login.component.ts
@@ -2,6 +2,7 @@ import { Component, ChangeDetectionStrategy, OnInit } from '@angular/core';
 import { AuthService } from '@blockframes/auth/+state';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { ActivatedRoute } from '@angular/router';
+import { MatSnackBar } from '@angular/material/snack-bar';
 
 @Component({
   selector: 'event-login',
@@ -13,7 +14,8 @@ export class EventLoginComponent implements OnInit {
 
   constructor(
     private service: AuthService,
-    private route: ActivatedRoute
+    private route: ActivatedRoute,
+    private snackBar: MatSnackBar,
     ) { }
   
   public eventId: string;
@@ -29,6 +31,10 @@ export class EventLoginComponent implements OnInit {
   }
 
   async validateLogin() {
+    if (!this.loginForm.valid) {
+      this.snackBar.open('Form invalid, please check error messages', 'close', { duration: 2000 });
+      return;
+    }
     await this.service.deleteAnonymousUser();
     await this.service.signin(this.loginForm.value.email, this.loginForm.value.password);
   }
